Redesign the `ARModel` around batched time series
The current timeseries + groups representation used in the ARModel
interface is somewhat inconvenient, since it requires forwarding the
groups argument down through the nested estimators when the model is
used in a pipeline. Instead, we will represent multiple time series as a
"batched time series", which is a list/array/series of individual time
series arrays. All sklearn utils including CV should work well with this
format. In addition, it is a natural intermediate representation when
the underlying data are stored in a data frame with a time series
column.